Add keep_md to make-sheets

+++ make-sheets.py
@@ -19,6 +19,7 @@ parser.add_argument('sheets_header', type=str,
 parser.add_argument('--team_names', type=str, nargs='?', default=None, help='Path to text file with team names (one per line)')
 parser.add_argument('--key', type=bool, nargs='?', default=False, help='Whether to print a key sheet')
 parser.add_argument('--output_file', type=str, nargs='?', default="answer_sheets.pdf", help='Filename of the answer sheets')
+parser.add_argument('--keep_md', type=bool, nargs='?', default=False, help='Whether to keep the Markdown file')
 parser.add_argument('--no_chain', type=bool, nargs='?', default=False, help='Whether to chain the output to Quarto immediately')
 args = parser.parse_args()
 
@@ -174,4 +175,5 @@ subprocess.run(["quarto",
                 "--to", "pdf",
                 "-o", args.output_file])
 
-os.remove(TEMP_FILENAME)
+if not args.keep_md:
+    os.remove(TEMP_FILENAME)
